COUR-307 | Add missing (WIP) modals to the TopBar, move modal context dispatcher types/functionality around
