+++ react/src/components/common/layout/TopBar/index.tsx
@@ -18,9 +18,12 @@ import ListItemAvatar from '@mui/material/ListItemAvatar'
 import Avatar from '@mui/material/Avatar'
 import Typography from '@mui/material/Typography'
 import { getNameInitials } from '@cfModule/utility/utilityFunctions'
+import CreateProgramModal from '@cfModule/components/common/dialog/CreateProgram'
+import CreateProjectModal from '@cfModule/components/common/dialog/CreateProject'
+import CreateCourseModal from '@cfModule/components/common/dialog/CreateCourse'
+import CreateActivityModal from '@cfModule/components/common/dialog/CreateActivity'
 import ResetPasswordModal from '@cfModule/components/common/dialog/ResetPassword'
-import { DIALOG_TYPE } from '@cfModule/components/common/dialog/context'
-import { useDialog } from '@cfModule/components/common/dialog'
+import { DIALOG_TYPE, useDialog } from '@cfModule/components/common/dialog'
 
 import {
   TopBarWrap,
@@ -32,12 +35,8 @@ import {
 import { getTargetProjectMenu } from '@XMLHTTP/API/workflow'
 import { TopBarProps } from '@cfModule/types/common'
 
-// TODO: Extract this into separate component/modal functionality
-// see https://course-flow.atlassian.net/browse/COUR-307
-// supported "add" menu actions
-export type CreateActionType = 'program' | 'activity' | 'course'
-
-export function openCreateActionModal(type: CreateActionType) {
+// TODO: clean up and move into create modals functionality
+function openCreateActionModal(type: 'program' | 'activity' | 'course') {
   const createUrl = COURSEFLOW_APP.config.create_path[type]
   COURSEFLOW_APP.tinyLoader.startLoad()
   getTargetProjectMenu<{ parentID: number }>(
@@ -93,7 +92,6 @@ const TopBar = ({ isTeacher, menus, notifications }: TopBarProps) => {
   const handleCreateClick = (resourceType: DIALOG_TYPE) => {
     closeAllMenus()
     dispatch(resourceType)
-    // openCreateActionModal(resourceType)
   }
 
   const addMenu = (
@@ -112,7 +110,7 @@ const TopBar = ({ isTeacher, menus, notifications }: TopBarProps) => {
       open={isAddMenuOpen}
       onClose={closeAllMenus}
     >
-      <MenuItem component="a" href={menus.add.projectUrl}>
+      <MenuItem onClick={() => handleCreateClick(DIALOG_TYPE.CREATE_PROJECT)}>
         {COURSEFLOW_APP.strings.project}
       </MenuItem>
       <MenuItem onClick={() => handleCreateClick(DIALOG_TYPE.CREATE_PROGRAM)}>
@@ -276,6 +274,10 @@ const TopBar = ({ isTeacher, menus, notifications }: TopBarProps) => {
       <ResetPasswordModal
         onSubmit={() => (window.location.href = menus.account.resetPasswordUrl)}
       />
+      <CreateProgramModal />
+      <CreateProjectModal />
+      <CreateCourseModal />
+      <CreateActivityModal />
     </TopBarWrap>
   )
 }
